backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.0.1
Flask-Cors==3.0.10
Werkzeug==2.0.1
PyJWT==2.1.0
cryptography==3.4.7
psycopg2-binary==2.9.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
transformers==4.8.2
torch==1.9.0
web3==5.23.1
requests==2.25.1
# Outras dependências necessárias
